requirements.txt
agno>=2.5.4
google-genai
python-dotenv
rich
